requirements_dev.txt
pytest>=8.0
pytest-asyncio>=0.23
pytest-homeassistant-custom-component>=0.13
ruff>=0.4
mypy>=1.10
homeassistant-stubs>=2025.1
